Start changing urls to use url_for.
This will be a nontrivial undertaking.  For every URL, we have to look up what function the url is actually calling and change to the url_for function.  We will also have to provide some sort of url mapping facility for the javascript code, since that code calls a lot of the urls directly.  We need to provide some sort of indirection for the javascript code since we might be running the code under a subdirectory of the server, so we at least need a prefix url.